fix:duplicate entries,each get removes previous data

+++ Project/crons.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from django.core.files.base import ContentFile
 import requests
 from datetime import date,datetime, timedelta,timezone
-from os.path import join
+from os  import path,remove
 import pandas as pd
 from geopy.geocoders import Nominatim
 #import classes from models
@@ -13,22 +13,23 @@ class GetWeekly(CronJobBase):
     schedule = Schedule(run_every_mins=720)
     code = "cron.GetWeekly"
 
-    def hour_elapsed(self):
-            latest_query = WorldData.objects.last()
-            days = (datetime.now(timezone.utc) - latest_query.time).days
-            seconds = (datetime.now(timezone.utc) - latest_query.time).seconds
-            return (days*12 + seconds//3600)
+    # def hour_elapsed(self):
+    #         latest_query = WorldData.objects.last()
+    #         days = (datetime.now(timezone.utc) - latest_query.time).days
+    #         seconds = (datetime.now(timezone.utc) - latest_query.time).seconds
+    #         return (days*12 + seconds//3600)
 
     def save_csv(self):
-        qs = WorldData.objects.order_by('-time').distinct()[:10]
-        listings = sorted(qs,key=lambda o:o.Mag)
-        listings.reverse()
+        listings = WorldData.objects.order_by('-Mag')[:10]
+        
         dict_list = [{'Magnitude':list.Mag,'latitude':list.lat,
         'longitude':list.lon,'country':list.country,'time':list.time} for list in listings]
         df = pd.DataFrame(dict_list)
         filestr = df.to_string()
-        # path = join(settings.MEDIA_ROOT, 'downloadable_csv', 'weekly_data.csv')
+        PATH= path.join(settings.MEDIA_ROOT, 'downloadable_csv', 'weekly_data.csv')
+        remove(PATH)
         WeeklyCsvFile.objects.all().delete()
+
         # with open(path,'w+') as f:
         #     f.write(filestr)
         weekly_data = WeeklyCsvFile()    
@@ -50,7 +51,7 @@ class GetWeekly(CronJobBase):
                 'mag',ascending=False).reset_index(
                     drop=True).head(10)
         obj_row = []
-
+        WorldData.objects.all().delete()
         for i in range(len(df)):
             lat = df.iloc[i]['latitude']
             lon = df.iloc[i]['longitude']
@@ -77,32 +78,32 @@ class GetWeekly(CronJobBase):
         WorldData.objects.bulk_create(obj_row)
 
     def do(self, *args, **options):
-        
-        
-        if self.hour_elapsed()>24:
-            end=datetime.now()
-            start = end-timedelta(days=7)
+        # if self.hour_elapsed()>24:
+        end=datetime.now()
+        start = end-timedelta(days=7)
 
-            data = requests.get(f'https://earthquake.usgs.gov/fdsnws/event/1/query?format=csv&starttime={start}&endtime={end}&minmagnitude=4.5')
-            #filepath = 'MapEarthQ/data.csv'
-            if data.status_code==200:
-                with open('data/data.csv', 'w') as writefile:
-                    writefile.write(data.text)
-                self.create_data()
-                self.save_csv()
+        data = requests.get(f'https://earthquake.usgs.gov/fdsnws/event/1/query?format=csv&starttime={start}&endtime={end}&minmagnitude=4.5')
+        #filepath = 'MapEarthQ/data.csv'
+        if data.status_code==200:
+            with open('data/data.csv', 'w') as writefile:
+                writefile.write(data.text)
+            self.create_data()
+            self.save_csv()
 
-            else:
-                print(data.status_code)
+        else:
+            print(data.status_code)
 
 class DeleteOldWeekly(CronJobBase):
     schedule = Schedule(run_monthly_on_days=1)
     code = "cron.DeleteOldWeekly"
 
     def do(self):
-        querylist = WorldData.objects.all()
-        for query in querylist:
-            days_passed = (datetime.now(timezone.utc) - query.time).days
-            if days_passed>=7:
-                print(f'{query.country} deleted!')
-                query.delete()
-                  
+        if WorldData.objects.exists():
+
+            querylist = WorldData.objects.all()
+            for query in querylist:
+                days_passed = (datetime.now(timezone.utc) - query.time).days
+                if days_passed>=7:
+                    print(f'{query.country} deleted!')
+                    query.delete()
+
